I fixed four problems.
The first two small  question is  change numDistis and 0 into numberOfDigits and 0.
Second one  is that i found first = square / modulus, so modulus shouldn't be 0, 1 will be better.
And when i run my program and i use calcultor, i found why 3 digits cannot be passed. Because i is not enough big, so i need to increase the numberOfDigits.